requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
google-generativeai==0.8.5
duckduckgo-search==7.5.3
